requirements.txt
opencv-python
numpy
requests
pyyaml
ultralytics